+++ core/src/Kokkos_Macros.hpp
@@ -55,7 +55,7 @@
 
 #ifndef KOKKOS_DONT_INCLUDE_CORE_CONFIG_H
 #include <KokkosCore_config.h>
-#include <Cuda/Kokkos_Cuda_NvidiaGpuArchitectures.hpp>
+#include <impl/Kokkos_NvidiaGpuArchitectures.hpp>
 #endif
 
 //----------------------------------------------------------------------------
